Getting familiar with FIR implementations and writing reference implementation to validate a more obscure, but possibly optimized, implementation that I have come across. The plan is to look at implementing a polyphase decimating filter: https://www.dsprelated.com/showarticle/198.php https://safe.nrao.edu/wiki/pub/CICADA/PolyPhasefilterbanks/polyphase.pdf# http://www.actel.com/ipdocs/CoreFIR_PF_HB.pdf
The decimating filter will initially be used to convert the acquisition domain (sample as fast as the ADC and hardware will allow) to the sample domain (selectable according to prefered resource usage, e.g. number of captured samples) in the hope that this will result in a more accurate representation of the signal - it might not and might be a waste of time, but hey, I've just learnt about polyphase filters.